routing/http: add support for GetClosestPeers (IPIP-476) (#1021)
This adds for GetClosestPeers into the routing packages. Boxo does not implement the actual GetClosestPeers call, but it provides the facilities to provide this endpoint in routing/v1 servers and clients.

GetClosestPeers finds the closest peers to a given DHT-Key. We only accept CIDs and Peer.IDs as keys (either as legacy Peer IDs or CID-formatted peer IDs). This helps setting guidelines for the intended usage of this method.

Co-authored-by: Marcin Rataj <lidel@lidel.org>
Co-authored-by: Andrew Gillis <11790789+gammazero@users.noreply.github.com>